src/python programs/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.2
distro==1.8.0
faiss-cpu==1.7.4
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
futures==3.0.5
googletrans2==2.3.0
goslate==1.5.4
greenlet==3.0.1
gTTS==2.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iso639==0.1.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.334
langdetect==1.0.9
langsmith==0.0.63
libretranslatepy==2.1.1
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
PyDictionary==2.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
translate==3.6.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
